Edited solution structure Models and Enums has moved to DepositUI.Core project
