--- DepositUI.BLL/DTOs/DepositCalcDTO.cs
@@ -1,11 +1,0 @@
-﻿namespace DepositUI.BLL.DTOs
-{
-    public class DepositCalcDTO
-    {
-        public int Month { get; set; }
-
-        public double PercentAdded { get; set; }
-
-        public double TotalAmount { get; set; }
-    }
-}
