+++ DepositUI.BLL/DepositUI.BLL.csproj
@@ -11,4 +11,8 @@
     <PackageReference Include="System.Net.Http.Json" Version="5.0.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\DepositUI.Core\DepositUI.Core.csproj" />
+  </ItemGroup>
+
 </Project>
