+++ DepositUI.BLL/Interfaces/IDepositService.cs
@@ -1,4 +1,4 @@
-﻿using DepositUI.BLL.DTOs;
+﻿using DepositUI.Core.Data;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -6,11 +6,11 @@ namespace DepositUI.BLL.Interfaces
 {
     public interface IDepositService
     {
-        public Task<List<DepositDTO>> GetDepositsAsync(int startIndex, int count);
+        public Task<List<DepositModel>> GetDepositsAsync(int startIndex, int count);
 
-        public Task<List<DepositCalcDTO>> GetDepositDetailsAsync(int depositId);
+        public Task<List<DepositCalc>> GetDepositDetailsAsync(int depositId);
 
-        public Task<List<DepositCalcDTO>> CalculateDepositAsync(DepositDTO deposit);
+        public Task<List<DepositCalc>> CalculateDepositAsync(DepositModel deposit);
 
         public Task<string> GetDepositCSV(int depositId);
     }
