+++ DepositUI/Components/DepositDetailsTable.razor
@@ -1,4 +1,4 @@
-﻿@using DepositUI.Data
+﻿@using DepositUI.Core.Data
 @inject IJSRuntime JS;
 <table class="result-table">
     <thead class="table-head">
