+++ DepositUI/DepositUI.csproj
@@ -6,8 +6,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="AutoMapper" Version="10.1.1" />
-    <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="8.1.1" />
     <PackageReference Include="Microsoft.AspNetCore.Components.Authorization" Version="3.1.19" />
     <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly" Version="3.2.1" />
     <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.Authentication" Version="3.2.1" />
@@ -20,6 +18,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\DepositUI.BLL\DepositUI.BLL.csproj" />
+    <ProjectReference Include="..\DepositUI.Core\DepositUI.Core.csproj" />
   </ItemGroup>
 
 </Project>
