+++ DepositUI/_Imports.razor
@@ -11,5 +11,6 @@
 @using DepositUI
 @using DepositUI.Shared
 @using DepositUI.Components;
-@using DepositUI.Enums;
+@using DepositUI.Core.Data;
+@using DepositUI.Core.Enums;
 @using System.Net.Http.Headers
